# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is workstation backend with rk-boot.",
        "title": "Workstation",
        "contact": {
            "name": "PointGoal team",
            "url": "https://github.com/pointgoal/workstation",
            "email": "lark@pointgoal.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/oauth/callback/github": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "Oauth callback",
                "operationId": "40",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
...
}